pyproject.toml
[project]
name = "graphrag"
version = "0.1.0"
description = "A proof of concept for building and querying knowledge graphs from markdown files"
requires-python = ">=3.9"
dependencies = [
    "llama-index-core>=0.10.1",
    "llama-index-llms-openai>=0.1.1",
    "llama-index-embeddings-openai>=0.1.1",
    "llama-index-readers-file>=0.1.1",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "networkx>=3.2.1",
    "matplotlib>=3.8.2",
    "streamlit>=1.29.0",
    "pyvis>=0.3.2",
    "watchdog==6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/graphrag"]

[project.scripts]
graphrag = "graphrag.graph_rag_demo:app"
